requirements.txt
python-dotenv
pyopenssl